refactor(admin/monitoring): 대시보드 구조 및 성능/UX 대대적 개선 기존 통합 API를 health/uptime/analytics/errors 등 개별 API로 분리
